% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paisaje-package.R
\docType{package}
\name{paisaje-package}
\alias{paisaje}
\alias{paisaje-package}
\title{paisaje: Spatial and Environmental Data Tools for Landscape Ecology}
\description{
Provides functions for landscape analysis and data retrieval. The package allows users to download environmental variables from global datasets (e.g., WorldClim, ESA WorldCover, Nighttime Lights), and to compute spatial and landscape metrics using a hexagonal grid system based on the H3 spatial index. It is useful for ecological modeling, biodiversity studies, and spatial data processing in landscape ecology. Fick and Hijmans (2017) \doi{10.1002/joc.5086}. Zanaga et al. (2022) \doi{10.5281/zenodo.7254221}. Uber Technologies Inc. (2022) "H3: Hexagonal hierarchical spatial index".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://manuelspinola.github.io/paisaje/}
  \item \url{https://github.com/ManuelSpinola/paisaje}
  \item Report bugs at \url{https://github.com/ManuelSpinola/paisaje/issues}
}

}
\author{
\strong{Maintainer}: Manuel Spínola \email{mspinola10@gmail.com} (\href{https://orcid.org/0000-0002-7839-1908}{ORCID})

}
\keyword{internal}
