% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_it_metrics.R
\name{calculate_it_metrics}
\alias{calculate_it_metrics}
\title{Calculate information theory landscape metrics}
\arguments{
\item{landscape_raster}{A categorical raster object: SpatRaster.}

\item{aoi_sf}{The spatial area of interest as an sf object.}
}
\value{
An sf object
}
\description{
Calculate information theory landscape metrics
}
\examples{

library(sf)
library(spDataLarge)

landscape_raster <- rast(system.file("raster/nlcd.tif",
package = "spDataLarge"))

bbox <- st_bbox(landscape_raster) |>
st_as_sfc(crs = EPSG:26912) |>
st_as_sf()

h3_bbox <- get_h3_grid(bbox, resolution = 6) |>
st_transform(26912)


result_sf <- calculate_it_metrics(landscape_raster, h3_bbox)

}
