% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cr_outline_c}
\alias{cr_outline_c}
\title{Costa Rica Continental Outline}
\format{
An `sf` object containing polygon geometry of Costa Rica.
}
\source{
Adapted from publicly available geographic data.
}
\usage{
cr_outline_c
}
\description{
A simplified outline of Costa Rica as an `sf` object.
}
\examples{
library(sf)
plot(cr_outline_c)
}
\keyword{datasets}
