% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cat_esa_10m.R
\name{create_cat_esa_10m}
\alias{create_cat_esa_10m}
\title{Create a categorical land cover raster from
 \href{https://esa-worldcover.org/en}{ESA WorldCover}}
\usage{
create_cat_esa_10m(input_raster, output_raster = NULL)
}
\arguments{
\item{input_raster}{Path to your input raster file, which
should be the original Copernicus land cover raster.}

\item{output_raster}{(optional): Path where the categorized
raster should be saved. If NULL, the raster is not saved to disk.}
}
\value{
A raster
}
\description{
Create a categorical land cover raster from
 a land cover raster from copernicus 10m resolution
}
\references{
WorldCover 2020 v100

Zanaga, D., Van De Kerchove, R., De Keersmaecker, W.,
Souverijns, N., Brockmann, C., Quast, R., Wevers, J.,
Grosu, A., Paccini, A., Vergnaud, S., Cartus, O., Santoro,
M., Fritz, S., Georgieva, I., Lesiv, M., Carter, S., Herold,
M., Li, Linlin, Tsendbazar, N.E., Ramoino, F., Arino, O.,
2021. ESA WorldCover 10 m 2020 v100.
https://doi.org/10.5281/zenodo.5571936


WorldCover 2021 v200

Zanaga, D., Van De Kerchove, R., Daems, D., De Keersmaecker,
 W., Brockmann, C., Kirches, G., Wevers, J., Cartus, O.,
 Santoro, M., Fritz, S., Lesiv, M., Herold, M., Tsendbazar,
 N.E., Xu, P., Ramoino, F., Arino, O., 2022. ESA WorldCover
 10 m 2021 v200. https://doi.org/10.5281/zenodo.7254221

ESA WorldCover project 2020 and 2021 / Contains modified
Copernicus Sentinel data (2020 and 2021) processed by ESA
WorldCover consortium
With year either 2020 or 2021 for the WorldCover 2020
and 2021 map, respectively.

\href{https://esa-worldcover.org/en}{ESA WorldCover}
}
