% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cat_esa_10m.R
\name{create_cat_esa_10m}
\alias{create_cat_esa_10m}
\title{Create Categorical Land Cover Raster from Copernicus
Land Cover Data \href{https://esa-worldcover.org/en}{ESA WorldCover}}
\usage{
create_cat_esa_10m(input_raster, output_raster = NULL)
}
\arguments{
\item{input_raster}{A character string specifying the file
path to the input raster (Copernicus land cover raster) that
is to be reclassified.}

\item{output_raster}{A character string specifying the file
path where the output categorical raster will be saved.
If `NULL`, the raster will not be saved.}
}
\value{
A `SpatRaster` object containing the reclassified
categorical land cover raster.
}
\description{
This function loads a Copernicus land cover raster,
reclassifies it into specified land cover categories, and
optionally saves the reclassified raster to a file. The
reclassification is based on predefined original classes and
their corresponding new categorical classes.
}
\details{
The function defines a mapping of original land cover classes
to new categorical classes. The reclassification is performed
using the `levels` and `coltab` functions from the `terra`
package. The reclassified raster will have colors associated
with each class as defined in the `coltb` data frame.
}
\examples{
\dontrun{
# Create categorical land cover raster
categorical_raster <- create_cat_esa_10m("path_to_input_raster.tif",
                                           "path_to_output_raster.tif")
}

}
\references{
WorldCover 2020 v100

Zanaga, D., Van De Kerchove, R., De Keersmaecker, W.,
Souverijns, N., Brockmann, C., Quast, R., Wevers, J.,
Grosu, A., Paccini, A., Vergnaud, S., Cartus, O., Santoro,
M., Fritz, S., Georgieva, I., Lesiv, M., Carter, S., Herold,
M., Li, Linlin, Tsendbazar, N.E., Ramoino, F., Arino, O.,
2021. ESA WorldCover 10 m 2020 v100.
https://doi.org/10.5281/zenodo.5571936


WorldCover 2021 v200

Zanaga, D., Van De Kerchove, R., Daems, D., De Keersmaecker,
 W., Brockmann, C., Kirches, G., Wevers, J., Cartus, O.,
 Santoro, M., Fritz, S., Lesiv, M., Herold, M., Tsendbazar,
 N.E., Xu, P., Ramoino, F., Arino, O., 2022. ESA WorldCover
 10 m 2021 v200. https://doi.org/10.5281/zenodo.7254221

ESA WorldCover project 2020 and 2021 / Contains modified
Copernicus Sentinel data (2020 and 2021) processed by ESA
WorldCover consortium
With year either 2020 or 2021 for the WorldCover 2020
and 2021 map, respectively.

\href{https://esa-worldcover.org/en}{ESA WorldCover}
}
