% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cat_esa_10m.R
\name{create_cat_esa_10m}
\alias{create_cat_esa_10m}
\title{Create Categorical Land Cover Raster from Copernicus
Land Cover Data \href{https://esa-worldcover.org/en}{ESA
WorldCover}}
\usage{
create_cat_esa_10m(land_cover)
}
\arguments{
\item{land_cover}{A SpatRaster object representing the input
land cover raster. The raster should contain land cover
classes defined by the Copernicus program.}
}
\value{
A SpatRaster object representing the reclassified
categorical land cover raster.
}
\description{
This function takes a `SpatRaster` object representing
Copernicus land cover data, reclassifies it into categorical
values based on specified land cover classes, and returns
the resulting categorical raster.
}
\details{
The function defines land cover categories and
assigns corresponding colors. It reclassifies the input
raster based on predefined categories.
}
\examples{
\dontrun{
# Assuming 'land_cover_raster' is a SpatRaster object
cat_raster <- create_cat_esa_10m(land_cover_raster)
}

}
\references{
WorldCover 2020 v100

Zanaga, D., Van De Kerchove, R., De Keersmaecker, W.,
Souverijns, N., Brockmann, C., Quast, R., Wevers, J.,
Grosu, A., Paccini, A., Vergnaud, S., Cartus, O., Santoro,
M., Fritz, S., Georgieva, I., Lesiv, M., Carter, S., Herold,
M., Li, Linlin, Tsendbazar, N.E., Ramoino, F., Arino, O.,
2021. ESA WorldCover 10 m 2020 v100.
https://doi.org/10.5281/zenodo.5571936


WorldCover 2021 v200

Zanaga, D., Van De Kerchove, R., Daems, D., De Keersmaecker,
 W., Brockmann, C., Kirches, G., Wevers, J., Cartus, O.,
 Santoro, M., Fritz, S., Lesiv, M., Herold, M., Tsendbazar,
 N.E., Xu, P., Ramoino, F., Arino, O., 2022. ESA WorldCover
 10 m 2021 v200. https://doi.org/10.5281/zenodo.7254221

ESA WorldCover project 2020 and 2021 / Contains modified
Copernicus Sentinel data (2020 and 2021) processed by ESA
WorldCover consortium
With year either 2020 or 2021 for the WorldCover 2020
and 2021 map, respectively.

\href{https://esa-worldcover.org/en}{ESA WorldCover}
}
