% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_num_raster.R
\name{extract_num_raster}
\alias{extract_num_raster}
\title{Extract raster values using weighted mean}
\usage{
extract_num_raster(spat_raster_multi, sf_hex_grid)
}
\arguments{
\item{num_raster}{A numeric SpatRaster}

\item{grid_sf}{An sf object with polygons}
}
\value{
An sf object with raster values added
}
\description{
Extracts numeric raster values for each polygon in `grid_sf`
  and computes a weighted mean using `exactextractr`.
}
\examples{
\dontrun{
library(terra)
library(sf)
library(exactextractr)

r <- rast(system.file("ex/elev.tif", package = "terra"))
poly <- st_as_sf(st_sfc(st_polygon(list(rbind(
  c(0,0), c(1,0), c(1,1), c(0,1), c(0,0)
))), crs = crs(r)))

result <- extract_num_raster(r, poly, fun = mean)
}
}
