% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_copernicus_10m.R
\name{get_copernicus_10m}
\alias{get_copernicus_10m}
\title{Download land cover data}
\usage{
get_copernicus_10m(aoi_sf, year = 2021, output_folder = ".")
}
\arguments{
\item{aoi_sf}{(\code{character}) an sf_object which defines
the area of interest (AOI). an sf object can represent any
geographical area, such as a country, state, or custom boundary.}

\item{year}{the Copernicus Global Land Service has different
versions of its land cover data for different years. As of now,
the available years are: 2020: This corresponds to the v100
product version, and 2021: This corresponds to the v200
product version.}

\item{output_folder}{specifies the directory where data files
will be saved. The default value for output_folder is ".",
which represents the current working directory.}
}
\description{
This function allows you to download land cover data from
 the \href{https://esa-worldcover.org/en}{ESA WorldCover}.
}
\examples{

library(sf)

nc = st_read(system.file("shape/nc.shp", package="sf"))

get_copernicus_10m(nc, year = 2021, output_folder = ".")


}
