% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_h3_grid.R
\name{get_h3_grid}
\alias{get_h3_grid}
\title{Generate an H3 Hexagonal Grid for an sf Object}
\usage{
get_h3_grid(sf_object, resolution = 7, expand_factor = 0.1)
}
\arguments{
\item{sf_object}{`sf` object. Must have a valid CRS.}

\item{resolution}{Integer. H3 resolution level (1–16). Default 7.}

\item{expand_factor}{Numeric. Expands bounding box to ensure coverage. Default 0.1.}
}
\description{
Generates a hexagonal grid of H3 cells at a specified resolution
that intersect with a given `sf` object. Wrapper for h3jsr functions.
}
\examples{
\dontrun{
library(sf)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
h3_grid_sf <- get_h3_grid(nc, resolution = 7)
}
}
