% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_h3_grid.R, R/get_nighlight_data.R
\name{get_h3_grid}
\alias{get_h3_grid}
\title{Create an H3 grid for a spatial object}
\usage{
get_h3_grid(sf_object, resolution)
}
\arguments{
\item{sf_object}{(\code{character}) An sf_object where to create the grid.}

\item{resolution}{(\code{numeric}) The option for the resolution of image to
download. Should between 1 and 16.}
}
\description{
This function allows you to create an H3 grid for a specific region.

This function is a wrapper of functions from the \href{https://obrl-soil.github.io/h3jsr/}{h3jsr} package.

This function allows you to create an H3 grid for a specific region.

This function is a wrapper of functions from the \href{https://obrl-soil.github.io/h3jsr/}{h3jsr} package.
}
\details{
\href{https://h3geo.org}{H3, a hexagonal hierarchical geospatial indexing system}
}
\examples{

Assuming you have an sf object called 'my_sf_object'
h3_grid_sf <- get_h3_grid(my_sf_object, resolution = 7)

}
\references{
O'Brien L (2023). h3jsr: Access Uber's H3 Library. R
package version 1.3.1,
<https://CRAN.R-project.org/package=h3jsr>.
}
