% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_h3_grid.R
\name{get_h3_grid}
\alias{get_h3_grid}
\title{Generate an H3 Hexagonal Grid for an sf Object}
\usage{
get_h3_grid(sf_object, resolution)
}
\arguments{
\item{sf_object}{`sf` object. The spatial object for which the
H3 grid will be generated. Must have a valid CRS.}

\item{resolution}{Integer. The H3 resolution level for the
hexagonal grid. Higher values create smaller hexagons.
Default is 7. It should between 1 and 16}

\item{expand_factor}{Numeric. A small value that expands
the bounding box of the `sf_object` to ensure complete
coverage by the hexagonal grid.
This helps avoid edge cases where the grid might not fully
cover the input object. Default is 0.001, which adjusts the
bounding box by this amount on all sides.}
}
\description{
This function generates a hexagonal grid of H3 cells at a
specified resolution level that intersect with a given `sf`
object.

This function is a wrapper of functions from
the \href{https://obrl-soil.github.io/h3jsr/}{h3jsr} package.
}
\details{
\href{https://h3geo.org}{H3, a hexagonal hierarchical geospatial indexing system}
}
\examples{

library(sf)

nc = st_read(system.file("shape/nc.shp", package="sf"))

h3_grid_sf <- get_h3_grid(nc, resolution = 7,
expand_factor = 0.1)


}
\references{
O'Brien L (2023). h3jsr: Access Uber's H3 Library. R
package version 1.3.1,
<https://CRAN.R-project.org/package=h3jsr>.
}
