% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nighlight_data.R
\name{get_nightlight_data}
\alias{get_nightlight_data}
\title{Download nightlight data}
\description{
This function allows you to download night lightdata from
 the \href{https://eogdata.mines.edu/}{The Earth Observation
  Group (EOG)}.
}
\examples{

get_nightlight_data(2024, 3, "v10", ".", timeout = 1200)


}
