% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_records.R
\name{get_records}
\alias{get_records}
\title{Retrieve species occurrence records within an Area of Interest}
\usage{
get_records(species, aoi_sf, providers = NULL,
limit = 500, remove_duplicates = FALSE, date = NULL)
}
\arguments{
\item{species}{A character vector of species names to query.}

\item{aoi_sf}{An `sf` object representing the area of interest.}

\item{providers}{A character vector of data providers to query (e.g., "gbif", "inat").}

\item{limit}{An integer specifying the maximum number of records to retrieve from each provider. Default is 500.}

\item{remove_duplicates}{A logical value indicating whether to remove duplicate geometries from the resulting `sf` object. Default is FALSE.}

\item{date}{A character vector of length 2 specifying the date range (e.g., c("YYYY-MM-DD", "YYYY-MM-DD")). Records outside this range will be excluded.}
}
\value{
An `sf` object containing the species occurrence records that fall within the specified AOI and meet the query criteria. Returns `NULL` if no records are found or if there are issues with the input.
}
\description{
This function retrieves species occurrence records from
specified data providers within a given area of interest (AOI).
The records are returned as an `sf` object, optionally with
duplicates removed based on geometry.
This function is a wrapper of the occ function from the
\href{https://CRAN.R-project.org/package=spocc}{spocc package}
}
\examples{
\dontrun{
nc <- sf::st_read(system.file("shape/nc.shp", package="sf"))
records <- get_records(species = "Lynx rufus", nc, providers = c("gbif", "inat"))
}
}
