% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_records.R
\name{get_records}
\alias{get_records}
\title{Download species occurrences from different sources}
\usage{
get_records(species_name, aoi_sf, providers = NULL)
}
\arguments{
\item{species_name}{Species name.}

\item{aoi_sf}{The spatial area of interest as an sf object.}

\item{providers}{The providers argument allows you to specify
multiple data sources (e.g., "gbif", "inat", etc.).}
}
\value{
An sf object
}
\description{
This function is wrapper of the occ function from the
 \href{https://cran.r-project.org/web/packages/spocc/index.html}{spocc package}.
}
\examples{

nc = sf::st_read(system.file("shape/nc.shp", package="sf"))

records <- get_records("Lynx rufus", nc, providers = c("gbif",
"inat"))


}
