% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_records.R
\name{get_records}
\alias{get_records}
\title{Retrieve Species Occurrence Records Within an Area
of Interest}
\usage{
get_records(species_name,
aoi_sf,
providers = NULL,
remove_duplicates = FALSE))
}
\arguments{
\item{species_name}{Character. The scientific name of the
species to search for.}

\item{aoi_sf}{`sf` object. The area of interest as a simple
features (sf) object. Must have a valid CRS.}

\item{providers}{Character vector. Data providers to query
for species occurrence data (e.g., `c("gbif", "inat")`).
See the `spocc` package documentation for available options.}

\item{remove_duplicates}{Logical. If `TRUE`, removes
duplicate records based on geometry to avoid counting the same
location multiple times. Default is `FALSE`.}
}
\value{
An `sf` object containing species occurrence records
that intersect with the AOI.

An `sf` object containing species occurrence records
that intersect with the AOI.
}
\description{
This function retrieves species occurrence records from
specified data providers within a given area of interest (AOI).
The records are returned as an `sf` object, optionally with
duplicates removed based on geometry.
This function is wrapper of the occ function from the
 \href{https://cran.r-project.org/web/packages/spocc/index.html}{spocc package}.
}
\examples{

nc = sf::st_read(system.file("shape/nc.shp", package="sf"))

records <- get_records("Lynx rufus", nc, providers = c("gbif",
"inat"))


}
