% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_records_by_hexagon.R
\name{get_records_by_hexagon}
\alias{get_records_by_hexagon}
\title{Download species records by hexagon of a hexagonal grid
of a specified resolution}
\usage{
get_records_by_hexagon(species_name, aoi_sf, res = 7, providers = NULL)
}
\arguments{
\item{species_name}{(\code{character}) Species name.}

\item{aoi_sf}{(\code{character}) The spatial area of interest as an sf object.}

\item{res}{(\code{numeric}) The option for the resolution of image to
download. Should between 1 and 16.}

\item{providers}{(\code{character}) The providers argument allows you to specify
multiple data sources (e.g., "gbif", "inat", etc.).}
}
\value{
An sf object
}
\description{
This function is wrapper of the occ function from the
 \href{https://cran.r-project.org/web/packages/spocc/index.html}{spocc package}.
}
\examples{


nc = sf::st_read(system.file("shape/nc.shp", package="sf"))

rec_hex <- get_records_by_hexagon("Lynx rufus", nc, res = 6,
providers = c("gbif", "inat"))



}
