% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_records_by_hexagon.R
\name{get_records_by_hexagon}
\alias{get_records_by_hexagon}
\title{Get Species Records by Hexagonal Grid}
\usage{
get_records_by_hexagon(species_name, aoi_sf, res = NULL,
providers = NULL, remove_duplicates = FALSE)
}
\arguments{
\item{species_name}{Character. The scientific name of the
species to search for.}

\item{aoi_sf}{`sf` object. The area of interest as a
simple features (sf) object. Must have a valid CRS.}

\item{res}{Integer. The H3 resolution level for the
hexagonal grid. Higher values create smaller hexagons.}

\item{providers}{Character vector. Data providers to
query for species occurrence data (e.g., `c("gbif", "inat")`).
See the `spocc` package documentation for options.}

\item{remove_duplicates}{Logical. If `TRUE`, removes
duplicate records based on geometry to avoid counting
the same location multiple times. Default is `FALSE`.}

\item{expand_factor}{Numeric. A small value that expands
the bounding box of the `sf_object` to ensure complete
coverage by the hexagonal grid.
This helps avoid edge cases where the grid might not fully
cover the input object. Default is 0.001, which adjusts the
bounding box by this amount on all sides.}
}
\value{
An `sf` object representing the hexagonal grid
with a `record_count` column indicating the number of
species occurrence records in each hexagon.
}
\description{
This function retrieves species occurrence data
within a specified area of interest (AOI) and aggregates
the records into hexagonal grid cells. The function uses
the H3 spatial index to generate hexagonal cells at a
specified resolution, counts the number of occurrence
records within each hexagon, and returns an `sf` object
with the record counts.
This function is wrapper of the occ function from the
 \href{https://cran.r-project.org/web/packages/spocc/index.html}{spocc package}.
}
\details{
The function performs the following steps:
1. Converts the area of interest to H3 cells at the specified resolution.
2. Queries species occurrence data within the bounding box of the AOI from specified providers.
3. Converts the species data into an `sf` object and optionally removes duplicate records based on geometry.
4. Counts the number of occurrence points that fall within each hexagon and adds this count to the hexagon grid.
}
\note{
Ensure that the AOI is provided as an `sf` object with a
valid CRS. The function automatically transforms the AOI to WGS84 (EPSG:4326) if necessary.
}
\examples{

nc = sf::st_read(system.file("shape/nc.shp", package="sf"))

rec_hex <- get_records_by_hexagon("Lynx rufus", nc, res = 6,
providers = c("gbif", "inat"), remove_duplicates = FALSE,
expand_factor = 0.1)



}
