% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_worldclim_historic.R
\name{get_worldclim_historic}
\alias{get_worldclim_historic}
\title{Download and process environmental variables made by worldclim
version 2.1.}
\usage{
get_worldclim_historic(
  var = "bio",
  res = 10,
  aoi = NULL,
  retries = 3,
  timeout = 300
)
}
\arguments{
\item{var}{A character string specifying the
climate variable to download. Possible values are:
.  \itemize{
    \item "tavg": Average temperature
    \item "tmin": Minimum temperature
    \item "tmax": Maximum temperature
    \item "prec": Precipitation
    \item "srad": Solar radiation
    \item "wind": Wind speed
    \item "vapr": Vapor pressure
    \item "bio": Bioclimatic variables
 }}

\item{res}{A numeric value indicating the resolution of the
data. Valid options are:
  \itemize{
    \item 0.5: 30 arc-seconds (~1 km)
    \item 2.5: 2.5 arc-minutes (~5 km)
    \item   5: 5 arc-minutes (~10 km)
    \item  10: 10 arc-minutes (~20 km)
 }}

\item{aoi}{A spatial object representing the area of
interest. Can be an `sf` or `terra` vector.}

\item{retries}{An integer specifying the number of times
to retry the download if it fails. Default is 3.}

\item{timeout}{A numeric value specifying the maximum
time (in seconds) to wait for a server response. Default is
300 seconds.}
}
\value{
A `SpatRaster` object containing the downloaded
 climate data, clipped to the specified AOI.
}
\description{
This function downloads historic worldclim version 2.1
and processes it according to the specified
parameters. It supports various climate variables and
resolutions, and can clip the downloaded data to a
specified area of interest (AOI). The function is
designed to handle download issues with retry logic
and customizable timeouts.
}
\examples{


nc <- st_read(system.file("shape/nc.shp", package="sf"))
nc <- st_transform(nc, crs = 4326)

climate_historic <- get_worldclim_historic(var =
 "tmin", res = 5, aoi = nc)

}
\references{
Fick, Stephen E., and Robert J.
Hijmans. "WorldClim 2: new 1-km spatial resolution climate
surfaces for global land areas." \emph{International journal
of climatology} 37.12 (2017): 4302-4315.\doi{10.1002/joc.5086}
}
