% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paisaje-package.R
\docType{package}
\name{paisaje-package}
\alias{paisaje}
\alias{paisaje-package}
\title{paisaje: Tools for landscape analysis}
\description{
This package contains functions to perform landscape analysis Some of the functions allow to get data for landscape analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://manuelspinola.github.io/paisaje/}
}

}
\author{
\strong{Maintainer}: Manuel Spínola \email{mspinola10@gmail.com} (\href{https://orcid.org/0000-0002-7839-1908}{ORCID})

}
\keyword{internal}
